pyproject.toml
[tool.poetry]
name = "goodreads_sync"
version = "0.1.0"
description = ""
authors = ["0xChloe <chloe@catastrophesoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "^6.0.11"
httpx = "^0.27.2"
qbittorrent-api = "^2024.8.65"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
